Renaming the base pahfit file for easier imports
